Rid of runtime singleton

+++ compiler/src/runtime/value.rs
@@ -7,29 +7,29 @@ use cranelift_codegen::ir::{InstBuilder, Value as CraneliftValue, types};
 use cranelift_frontend::FunctionBuilder;
 
 /// Tag bits for runtime type identification
-pub const TAG_NIL: i64 = 0;
-pub const TAG_BOOLEAN: i64 = 1;
-pub const TAG_NUMBER: i64 = 2;
-pub const TAG_OBJECT_REF: i64 = 3;
-pub const TAG_STRING_REF: i64 = 4;
-pub const TAG_MESSAGE_REF: i64 = 5;
+pub const TAG_NIL: u64 = 0;
+pub const TAG_BOOLEAN: u64 = 1;
+pub const TAG_NUMBER: u64 = 2;
+pub const TAG_OBJECT_REF: u64 = 3;
+pub const TAG_STRING_REF: u64 = 4;
+pub const TAG_MESSAGE_REF: u64 = 5;
 
 /// Number of bits used for the tag
-pub const TAG_BITS: i64 = 3;
-pub const TAG_MASK: i64 = (1 << TAG_BITS) - 1;
+pub const TAG_BITS: u64 = 3;
+pub const TAG_MASK: u64 = (1 << TAG_BITS) - 1;
 
 /// Encode a nil value in Cranelift IR
 pub fn encode_nil(builder: &mut FunctionBuilder) -> CraneliftValue {
     // Nil is just the tag with no payload
-    builder.ins().iconst(types::I64, TAG_NIL)
+    builder.ins().iconst(types::I64, TAG_NIL as i64)
 }
 
 /// Encode a boolean value in Cranelift IR
 pub fn encode_boolean(builder: &mut FunctionBuilder, value: bool) -> CraneliftValue {
     // Boolean is the tag plus 1 bit for true/false
     let bool_val = builder.ins().iconst(types::I64, if value { 1 } else { 0 });
-    let shifted = builder.ins().ishl_imm(bool_val, TAG_BITS);
-    let tag = builder.ins().iconst(types::I64, TAG_BOOLEAN);
+    let shifted = builder.ins().ishl_imm(bool_val, TAG_BITS as i64);
+    let tag = builder.ins().iconst(types::I64, TAG_BOOLEAN as i64);
     builder.ins().bor(shifted, tag)
 }
 
@@ -46,24 +46,24 @@ pub fn encode_number(builder: &mut FunctionBuilder, value: CraneliftValue) -> Cr
 pub fn encode_reference(
     builder: &mut FunctionBuilder,
     ptr: CraneliftValue,
-    tag: i64,
+    tag: u64,
 ) -> CraneliftValue {
     // Reference is the pointer shifted left by TAG_BITS, then OR'd with the tag
-    let shifted = builder.ins().ishl_imm(ptr, TAG_BITS);
-    let tag_val = builder.ins().iconst(types::I64, tag);
+    let shifted = builder.ins().ishl_imm(ptr, TAG_BITS as i64);
+    let tag_val = builder.ins().iconst(types::I64, tag as i64);
     builder.ins().bor(shifted, tag_val)
 }
 
 /// Decode a tagged value to extract the tag
 pub fn decode_tag(builder: &mut FunctionBuilder, value: CraneliftValue) -> CraneliftValue {
-    let mask = builder.ins().iconst(types::I64, TAG_MASK);
+    let mask = builder.ins().iconst(types::I64, TAG_MASK as i64);
     builder.ins().band(value, mask)
 }
 
 /// Check if a value has a specific tag
-pub fn has_tag(builder: &mut FunctionBuilder, value: CraneliftValue, tag: i64) -> CraneliftValue {
+pub fn has_tag(builder: &mut FunctionBuilder, value: CraneliftValue, tag: u64) -> CraneliftValue {
     let tag_value = decode_tag(builder, value);
-    let tag_const = builder.ins().iconst(types::I64, tag);
+    let tag_const = builder.ins().iconst(types::I64, tag as i64);
     builder.ins().icmp(
         cranelift_codegen::ir::condcodes::IntCC::Equal,
         tag_value,
@@ -73,7 +73,7 @@ pub fn has_tag(builder: &mut FunctionBuilder, value: CraneliftValue, tag: i64) -
 
 /// Decode a reference value to extract the pointer
 pub fn decode_reference(builder: &mut FunctionBuilder, value: CraneliftValue) -> CraneliftValue {
-    builder.ins().ushr_imm(value, TAG_BITS)
+    builder.ins().ushr_imm(value, TAG_BITS as i64)
 }
 
 /// Allocate a number object in the runtime heap
